@@ -3,11 +3,11 @@ import 'package:quick_summary/utils/consts.dart';
 
 class SaveButton extends StatelessWidget {
   const SaveButton({
-    Key key,
-    @required this.saveAction,
+    Key? key,
+    required this.saveAction,
   }) : super(key: key);
 
-  final Function() saveAction;
+  final Function()? saveAction;
 
   @override
   Widget build(BuildContext context) {
@@ -15,7 +15,7 @@ class SaveButton extends StatelessWidget {
       width: double.infinity,
       child: Padding(
         padding: EdgeInsets.symmetric(vertical: 0, horizontal: 2),
-        child: RaisedButton(
+        child: ElevatedButton(
           child: Text(Consts.saveTitle),
           onPressed: saveAction,
         ),
